+++ QSC/QSCTest/nistrng.h
@@ -2,13 +2,12 @@
 #define QSCTEST_NISTRNG_H
 /**
 * \file nistrng.h
-* \brief This is not a secure rng, and should ber used for testing purposes only.
+* \brief This is not a secure RNG, and should be used for testing purposes only.
 *
 * \endcode
 */
 
-#include <stdbool.h>
-#include <stdint.h>
+#include "../QSC/common.h"
 
 #define QSCTEST_NIST_RNG_SEED_SIZE 48
 #define QSCTEST_NIST_RNG_SUCCESS 0
@@ -16,16 +15,16 @@
 #define QSCTEST_NIST_RNG_BAD_OUTBUF -2
 #define QSCTEST_NIST_RNG_BAD_REQ_LEN -3
 
-typedef struct 
+typedef struct
 {
     uint8_t state[16];
-    size_t bpos;
-    size_t rmdr;
+    uint32_t bpos;
+    uint32_t rmdr;
     uint8_t key[32];
     uint8_t ctr[16];
 } qsctest_nist_rng_state;
 
-typedef struct 
+typedef struct
 {
     uint8_t key[32];
     uint8_t ctr[16];
@@ -33,7 +32,7 @@ typedef struct
 } qsctest_nist_aes256_state;
 
  /**
- * \brief Initialize a user supplied kdf state instance
+ * \brief Initialize a user supplied KDF state instance
  *
  * \param ctx stores the current state of an instance of the seed expander
  * \param seed a 32 byte random value
@@ -54,7 +53,7 @@ int32_t qsctest_nistrng_kdf_initialize(qsctest_nist_rng_state* ctx, const uint8_
 int32_t qsctest_nistrng_kdf_generate(qsctest_nist_rng_state* ctx, uint8_t* output, size_t outlen);
 
 /**
-* \brief Initialize the random provider state with a seed 
+* \brief Initialize the random provider state with a seed
 * and optional personalization string
 *
 * \param seed 48 bytes of random seed
@@ -77,8 +76,8 @@ bool qsctest_nistrng_prng_generate(uint8_t* output, size_t outlen);
 /**
 * \brief Update the random provider with new keying material
 *
-* \param key the drbg key
-* \param counter the drbg counter
+* \param key the DRBG key
+* \param counter the DRBG counter
 * \param info the optional personalization string
 * \param infolen the length of the personalization string, can not exceed 48 bytes
 */
