+++ QSC/QSCTest/timer.c
@@ -15,7 +15,7 @@ uint64_t qsctest_timer_elapsed(clock_t start)
 	uint64_t msec;
 
 	diff = clock() - start;
-	msec = (diff * 1000) / CLOCKS_PER_SEC;
+	msec = (uint64_t)(diff * 1000) / CLOCKS_PER_SEC;
 
 	return msec;
 }
